binance-data-downloader/requirements.txt
python-binance
streamlit
tqdm